add first MVC Editor classes, not filled out yuet
